refactor(dashboard): split logic and UI for Order and Product pages for better maintainability
