+++ .env
@@ -1,6 +1,6 @@
-NEXT_PUBLIC_API_URL=https://byness-backend.vercel.app/api
-#NEXT_PUBLIC_API_URL=http://localhost:8000
-NEXT_PUBLIC_LOCAL_URL=https://byness.vercel.app
-#NEXT_PUBLIC_LOCAL_URL=http://localhost:3000
+#NEXT_PUBLIC_API_URL=https://byness-backend.vercel.app/api
+NEXT_PUBLIC_API_URL=http://localhost:8000
+#NEXT_PUBLIC_LOCAL_URL=https://byness.vercel.app
+NEXT_PUBLIC_LOCAL_URL=http://localhost:3000
 UPSTASH_REDIS_REST_URL=https://apparent-tadpole-55888.upstash.io
 UPSTASH_REDIS_REST_TOKEN=AdpQAAIjcDFlYTA0OTRjYmFlMjM0M2U5OTFkZDIxN2Y1Y2QyMGFiNXAxMA
