+++ src/components/Dashboard/Orders/LoadingSpinner.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+export function LoadingSpinner() {
+  return (
+    <div className="flex items-center justify-center h-64">
+      <div className="animate-spin rounded-full h-32 w-32 border-b-2 border-blue-500"></div>
+    </div>
+  );
+}
