+++ src/components/Dashboard/Products/utils/form/formHelpers.tsx
@@ -1,0 +1,21 @@
+'use client';
+
+/**
+ * Utility functions for form operations
+ */
+
+export const predefinedColors = [
+  '#000000', '#FFFFFF', '#FF0000', '#00FF00', '#0000FF',
+  '#FFFF00', '#FF00FF', '#00FFFF', '#FFA500', '#800080'
+];
+
+export const predefinedSizes = ['XS', 'S', 'M', 'L', 'XL', 'XXL', 'One Size'];
+
+export const generateSlug = (name: string): string => {
+  return name.toLowerCase().replace(/\s+/g, '-').replace(/[^a-z0-9-]/g, '');
+};
+
+export const calculateDiscountPercentage = (originalPrice: number, currentPrice: number): number => {
+  if (!originalPrice || originalPrice <= currentPrice) return 0;
+  return Math.round(((originalPrice - currentPrice) / originalPrice) * 100);
+};
